rl3

+++ rl/rl_loss2.py
@@ -79,10 +79,10 @@ def greedy_sample_with_logits(sls, els):
     els: end logits
     """
     max_seq_len = tf.shape(sls)[1]
-    start_sample = tf.multinomial(sls, 1, dtype=tf.int32)
+    start_sample = tf.cast(tf.multinomial(sls, 1), tf.int32)
     sps_mask = tf.sequence_mask(tf.squeeze(start_sample) - 1, maxlen=max_seq_len, dtype=tf.float32)  # start end 是可以重复的
     els = els * (1 - sps_mask) - 1e30 * sps_mask
-    end_sample = tf.multinomial(els, 1, dtype=tf.int32)
+    end_sample = tf.cast(tf.multinomial(els, 1), tf.int32)
 
     return start_sample, end_sample
